added new globalNavBar, removed the userMenu from dashboard and modified it on home. removed arrowBackIcon used to close dashboard and created animation instead
