+++ frontend/src/components/UserMenu/index.js
@@ -3,53 +3,40 @@ import { useDispatch, useSelector } from 'react-redux'
 import { push } from 'connected-react-router'
 import { makeStyles } from '@material-ui/core/styles'
 import {
-    Popover,
-    Avatar,
     Box,
     ListItem,
     ListItemText,
-    Divider,
     Grid,
 } from '@material-ui/core'
 import * as AuthSelectors from 'redux/auth/selectors'
-import MenuIcon from '@material-ui/icons/Menu'
 import JunctionTheme from 'junctionTheme.js'
-import LanguageMenu from 'components/LanguageMenu'
+import Button from 'components/generic/Button'
+import UserAvatar from 'components/UserAvatar'
 import { useMyProfilePreview } from 'graphql/queries/userProfile'
 
 import { useTranslation } from 'react-i18next'
 
 const useStyles = makeStyles(theme => ({
-    menuDot: {
-        width: '8px',
-        height: '8px',
-        margin: '2px',
-        borderRadius: '2px',
-        backgroundColor: 'rgba(0,0,0,0.3)',
-    },
     avatar: {
         marginLeft: '16px',
     },
-    hamburger: {
-        transition: theme.transitions.create(['transform'], {
-            duration: theme.transitions.duration.short,
-        }),
-    },
-    hamburgerOpen: {
-        transform: 'rotate(-90deg)',
-    },
-    hamburgerClosed: {
-        transform: 'rotate(0)',
-    },
-    tableBottom: {
-        borderBottom: 'none',
-    },
-    popover: {
-        borderRadius: '15px',
-    },
     menuBox: {
-        width: '220px',
-        borderRadius: '15px',
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
+        padding: '5px',
+        margin: '30px',
+        backgroundColor: 'white',
+        border: 'solid',
+        borderColor: theme.palette.primary.main,
+        '&:hover': {
+            backgroundColor: theme.palette.primary.light,
+        },
+    },
+    menuItem: {
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
     },
 }))
 
@@ -60,198 +47,63 @@ export default () => {
     const userId = idTokenPayload?.sub
     const [profile] = useMyProfilePreview()
     const dispatch = useDispatch()
-    const hasSuperAdmin = useSelector(AuthSelectors.hasSuperAdmin)
-    const hasRecruiterAccess = useSelector(AuthSelectors.hasRecruiterAccess)
     const classes = useStyles()
-    const [anchorEl, setAnchorEl] = useState(null)
-    const color = JunctionTheme.palette
 
-    const handleMenuOpen = e => {
-        setAnchorEl(e.currentTarget)
-    }
 
-    const handleMenuClose = () => {
-        setAnchorEl(null)
-    }
 
     if (!userId) {
         return (
             <Box display="flex" flexDirection="row" alignItems="center">
-                <MenuIcon
-                    fontSize="large"
-                    onClick={handleMenuOpen}
-                    className={`${classes.hamburger} ${anchorEl
-                        ? classes.hamburgerOpen
-                        : classes.hamburgerClosed
-                        }`}
-                />
 
-                <Popover
-                    open={Boolean(anchorEl)}
-                    onClose={handleMenuClose}
-                    anchorEl={anchorEl}
-                    anchorOrigin={{
-                        vertical: 'bottom',
-                        horizontal: 'right',
-                    }}
-                    transformOrigin={{
-                        vertical: 'top',
-                        horizontal: 'right',
-                    }}
-                    classes={{
-                        paper: classes.menuBox,
-                    }}
-                >
-                    <Box className={classes.menuBox}>
-                        <Grid container spacing={0}>
-                            <Grid item xs={12}>
-                                <ListItem button>
-                                    <ListItemText
-                                        primary={t('Sign_in_')}
-                                        onClick={() => dispatch(push('/login'))}
-                                    />
-                                </ListItem>
-                            </Grid>
-                            <Grid item xs={12}>
-                                <Divider variant="middle" />
-                            </Grid>
-                            <Grid item xs={6}>
-                                <LanguageMenu />
-                            </Grid>
-                        </Grid>
-                    </Box>
-                </Popover>
+
+                <Grid container spacing={0}>
+                    <Grid item xs={12}>
+                        <Button
+                            onClick={() => dispatch(push('/login'))}
+                            strong={true}
+                            className={classes.menuBox}
+                        >
+                            {t('Sign_in_')}
+                        </Button>
+                    </Grid>
+                </Grid>
+
+
             </Box>
         )
     }
 
-    const renderEventItems = () => {
-        //const items = [];
-        // TODO: Add links to event dashboard here for ongoing events
-        return null
-    }
 
-    const renderOtherItems = () => {
-        const items = []
-
-        // if (hasOrganiserAccess) {
-        items.push({
-            label: 'Create Event',
-            onClick: () => dispatch(push('/organise')),
-        })
-        // }
-
-        if (hasRecruiterAccess) {
-            items.push({
-                label: 'Recruitment',
-                onClick: () => dispatch(push('/recruitment')),
-            })
-        }
-
-        if (hasSuperAdmin) {
-            items.push({
-                label: 'Admin',
-                onClick: () => dispatch(push('/admin')),
-            })
-        }
-        // Grid Item isn't neccessary here, but put it in for consistency
-        if (items.length > 0) {
-            return (
-                <>
-                    {items.map(({ label, onClick }) => (
-                        <Grid key={label} item xs={12}>
-                            <ListItem button onClick={onClick}>
-                                <ListItemText primary={label} />
-                            </ListItem>
-                        </Grid>
-                    ))}
-                </>
-            )
-        }
-
-        return null
-    }
 
     return (
         <Box display="flex" flexDirection="row" alignItems="center">
-            <MenuIcon
-                fontSize="large"
-                onClick={handleMenuOpen}
-                className={`${classes.hamburger} ${anchorEl ? classes.hamburgerOpen : classes.hamburgerClosed
-                    }`}
-            />
 
-            <Popover
-                open={Boolean(anchorEl)}
-                onClose={handleMenuClose}
-                anchorEl={anchorEl}
-                anchorOrigin={{
-                    vertical: 'bottom',
-                    horizontal: 'right',
-                }}
-                transformOrigin={{
-                    vertical: 'top',
-                    horizontal: 'right',
-                }}
-                classes={{
-                    paper: classes.menuBox,
-                }}
-            >
-                <div className='tw-border-8 tw-border-white'>
-                    <Box className={classes.menuBox}>
-                        <Grid container spacing={0}>
-                            <Grid item xs={6}>
-                                <ListItem button>
-                                    <ListItemText
-                                        primary={t('Dashboard_')}
-                                        onClick={() => dispatch(push('/account'))}
-                                    />
-                                </ListItem>
-                            </Grid>
-                            <Grid item xs={6}>
-                                <ListItem
-                                    button
-                                    onClick={() => dispatch(push('/logout'))}
-                                >
-                                    <ListItemText primary={t('Log_out_')} />
-                                </ListItem>
-                            </Grid>
-                            <Grid item xs={12}>
-                                <ListItem button>
-                                    <ListItemText
-                                        primary={t('Edit_profile_')}
-                                        onClick={() =>
-                                            dispatch(push('/account/profile'))
-                                        }
-                                    />
-                                </ListItem>
-                            </Grid>
-                            {renderEventItems()}
-                            {renderOtherItems()}
+            <Grid container spacing={2} wrap="nowrap" >
+                <Grid item xs={4}>
 
-                            <Grid item xs={12}>
-                                <Divider variant="middle" />
-                            </Grid>
-                            <Grid item xs={6}>
-                                <LanguageMenu />
-                            </Grid>
-                        </Grid>
-                    </Box>
-                </div>
-            </Popover>
-            <div className='tw-ml-16 tw-rounded-full tw-border-8 tw-border-white'>
-                <Avatar
-                    className='tw-ml-16 tw-rounded-full tw-border-8 tw-border-white'//TODO: what is wrong with the border
-                    src={profile?.avatar}
-                    alt="Avatar"
-                    style={{
-                        width: '50px',
-                        height: '50px',
-                        border: `2px solid ${color['primary'].main}`,
-                        borderRadius: '50%',
-                    }}
-                />
-            </div>
-        </Box>
+                    <Button
+                        onClick={() => dispatch(push('/dashboard/default'))}
+                        strong={true}
+                        className={classes.menuBox}
+                    >
+
+                        {t('Dashboard_')}
+                    </Button>
+                </Grid>
+                <Grid item xs={5} wrap='nowrap'>
+
+                    <Button
+                        onClick={() => dispatch(push('/logout'))}
+                        strong={true}
+                        className={classes.menuBox}
+                    >
+                        {t('Log_out_')}
+                    </Button>
+                </Grid>
+                <Grid item xs={3} className={classes.menuItem}>
+                    <UserAvatar />
+                </Grid>
+            </Grid>
+        </Box >
     )
 }
