+++ frontend/src/components/navbars/GlobalNavBar/index.js
@@ -2,14 +2,17 @@ import React from 'react'
 
 import { makeStyles } from '@material-ui/core/styles'
 import UserMenu from 'components/UserMenu'
+import Button from 'components/generic/Button'
 
 import config from 'constants/config'
+import PlatformLogo from 'assets/logos/JO_wordmark_black.png'
+import WavePattern from 'assets/images/nawbar_waves.svg'
 
 const useStyles = makeStyles(theme => ({
     wrapper: {
         width: '100%',
         height: '60px',
-        background: 'white',
+        background: WavePattern,
         padding: theme.spacing(0, 2),
     },
     inner: {
@@ -21,23 +24,30 @@ const useStyles = makeStyles(theme => ({
         height: '100%',
     },
     wordmark: {
-        height: '20px',
+        height: '70px',
+        paddingLeft: '20px',
+        paddingTop: '5px',
+        paddingBottom: '5px',
+        backgroundColor: 'white',
     },
 }))
 
 export default () => {
     const classes = useStyles()
     return (
-        <div className={classes.wrapper}>
-            <div className={classes.inner}>
-                <a href="/home">
-                    <img
-                        src={config.LOGO_DARK_URL}
-                        className={classes.wordmark}
-                        alt={config.PLATFORM_OWNER_NAME + ' logo'}
-                    />
-                </a>
-                <UserMenu />
+        <div className='tw-w-full tw-py-2 tw-bg-wave-pattern' >
+            <div className={classes.wrapper}>
+                <div className={classes.inner}>
+
+                    <a href="/home">
+                        <img
+                            src={PlatformLogo}
+                            className={classes.wordmark}
+                            alt={config.PLATFORM_OWNER_NAME + ' logo'}
+                        />
+                    </a>
+                    <UserMenu />
+                </div>
             </div>
         </div>
     )
